working version that successfully emits logs on platforms other than windows
