+++ NLog.Targets.MauiLog.sln
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{76
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MauiApp2", "examples\MauiApp2\MauiApp2.csproj", "{F5DDCB81-5475-45A8-84B9-2F9880965288}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Targets.MauiLog.X53R", "NLog.Targets.MauiLog.X53R\NLog.Targets.MauiLog.X53R.csproj", "{19E01588-3686-4522-A126-E9FC5EA15644}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,8 +30,14 @@ Global
 		{F7C81AB8-6833-449D-80E3-68CEC129E692}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F5DDCB81-5475-45A8-84B9-2F9880965288}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F5DDCB81-5475-45A8-84B9-2F9880965288}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F5DDCB81-5475-45A8-84B9-2F9880965288}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
 		{F5DDCB81-5475-45A8-84B9-2F9880965288}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F5DDCB81-5475-45A8-84B9-2F9880965288}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F5DDCB81-5475-45A8-84B9-2F9880965288}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{19E01588-3686-4522-A126-E9FC5EA15644}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19E01588-3686-4522-A126-E9FC5EA15644}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19E01588-3686-4522-A126-E9FC5EA15644}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19E01588-3686-4522-A126-E9FC5EA15644}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
