+++ examples/MauiApp2/ViewModel/MainViewModel.cs
@@ -2,16 +2,20 @@
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
+using Microsoft.Extensions.Logging;
 
 namespace MauiApp2.ViewModel;
 
 public partial class MainViewModel : ObservableObject
 {
     IConnectivity connectivity;
-    public MainViewModel(IConnectivity connectivity)
+    private readonly ILogger<MainViewModel> _logger;
+
+    public MainViewModel(IConnectivity connectivity, ILogger<MainViewModel> logger)
     {
         Items = new ObservableCollection<string>();
         this.connectivity = connectivity;
+        _logger = logger;
     }
 
     [ObservableProperty]
@@ -49,6 +53,12 @@ public partial class MainViewModel : ObservableObject
     [RelayCommand]
     async Task Tap(string s)
     {
+        _logger.LogTrace("Trace");
+        _logger.LogDebug("Debug");
+        _logger.LogInformation("Info");
+        _logger.LogWarning("Warn");
+        _logger.LogError("Error");
+        _logger.LogCritical("Critical");
         await Shell.Current.GoToAsync($"{nameof(DetailPage)}?Text={s}");
     }
