+++ global.json
@@ -1,10 +1,6 @@
 {
   "sdk": {
-    "version": "6.0.400",
+    "version": "7.0.*",
     "rollForward": "latestMinor"
-  },
-  "msbuild-sdks": {
-    "MSBuild.Sdk.Extras": "3.0.44",
-    "Microsoft.Build.NoTargets": "3.3.0"
   }
 }
