+++ src/NLog.Targets.MauiLog/Config/SetupBuilderExtensions.cs
@@ -1,34 +1,34 @@
 ﻿using System;
 using NLog.Config;
+using NLog.Targets;
 
-namespace NLog
+namespace NLog;
+
+/// <summary>
+/// Extension methods to setup LogFactory options
+/// </summary>
+public static class SetupBuilderExtensions
 {
     /// <summary>
-    /// Extension methods to setup LogFactory options
+    /// Register the NLog.Web LayoutRenderers before loading NLog config
     /// </summary>
-    public static class SetupBuilderExtensions
+    public static ISetupBuilder RegisterMauiLog(this ISetupBuilder setupBuilder)
     {
-        /// <summary>
-        /// Register the NLog.Web LayoutRenderers before loading NLog config
-        /// </summary>
-        public static ISetupBuilder RegisterMauiLog(this ISetupBuilder setupBuilder)
-        {
-            setupBuilder.SetupExtensions(e => e.RegisterMauiLog());
-            return setupBuilder;
-        }
+        setupBuilder.SetupExtensions(e => e.RegisterMauiLog());
+        return setupBuilder;
+    }
 
-        /// <summary>
-        /// Register the NLog.Web LayoutRenderers before loading NLog config
-        /// </summary>
-        public static ISetupBuilder RegisterMauiLog(this ISetupBuilder setupBuilder, UnhandledExceptionEventHandler unhandledException)
-        {
-            if (unhandledException is null)
-                throw new ArgumentNullException(nameof(unhandledException));
+    /// <summary>
+    /// Register the NLog.Web LayoutRenderers before loading NLog config
+    /// </summary>
+    public static ISetupBuilder RegisterMauiLog(this ISetupBuilder setupBuilder, UnhandledExceptionEventHandler unhandledException)
+    {
+        if (unhandledException is null)
+            throw new ArgumentNullException(nameof(unhandledException));
 
-            setupBuilder.SetupExtensions(e => e.RegisterMauiLog());
-            NLog.Targets.MauiLog.MauiExceptions.UnhandledException -= unhandledException;   // Avoid double registration
-            NLog.Targets.MauiLog.MauiExceptions.UnhandledException += unhandledException;
-            return setupBuilder;
-        }
+        setupBuilder.SetupExtensions(e => e.RegisterMauiLog());
+        MauiExceptions.UnhandledException -= unhandledException;   // Avoid double registration
+        MauiExceptions.UnhandledException += unhandledException;
+        return setupBuilder;
     }
 }
